Added speed Increase
1. Speed will increase after a dish is made
2. removed coin cript and use ingredient controller instead
