+++ Assets/Scenes/GameplayScene.unity
@@ -484,7 +484,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 356856036}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8a8e676a6d2e647ab99ceba7ea0ae023, type: 3}
+  m_Script: {fileID: 11500000, guid: def1f0937341a4c9f89a333be034a0af, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   Ball: {fileID: 356856036}
