+++ Assets/Scripts/Score/ItemCollection.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8a8e676a6d2e647ab99ceba7ea0ae023
+guid: def1f0937341a4c9f89a333be034a0af
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
