--- Assets/Scripts/Tools/CoinScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2c3f4d4e159fd4eef8309858a4d9e6d0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
